multiple channels, autoreconnect, clickable channels, fixes
Added ability to register multiple channels to connect to (with
password). Added an autoreconnect. Some fixes to color code translation.
Added ability to click channel names to set that as default. The
"channel" option now contains the channel where messages will be sent.
Added more quit messages. Messages starting with "/" won't be sent to
IRC even when irc-only is true.